requirements.txt
customtkinter
pillow
torch
torchvision